changed transmit(u8* buf, u8 len) paradigm to better handle length for devs.   if len == 0:     uses the first byte of buffer as the length byte   if len != 0:     uses the value in length and *assumes no length byte in buffer!* the details of the radio being configured for variable or fixed length are handled for you.

+++ firmware/include/cc1111rf.h
@@ -68,7 +68,7 @@ int waitRSSI(void);
 void RxMode(void);
 void IdleMode(void);
 
-u8 transmit(__xdata u8*, u16 len);
+u8 transmit(__xdata u8*, u8 len);
 void appInitRf(void);   // in application.c
 void startRX(void);
 void stopRX(void);
